26.2.07OpenClaw某站视频转学习笔记并朗读、pdf转视频Skills/bilibili-study/requirements.txt
yt-dlp>=2024.1
requests>=2.31.0
webvtt-py>=0.4.3
opencc>=1.1.7
openai>=1.0.0
python-dotenv>=1.0.0
